+++ build.js
@@ -1,12 +1,22 @@
 const fs = require('fs');
+const path = require('path');
 const log = require('@vladmandic/pilogger');
 const Build = require('@vladmandic/build').Build;
 const APIExtractor = require('@microsoft/api-extractor');
 const tf = require('@tensorflow/tfjs-node');
 const package = require('./package.json');
 
-const modelsDir = '../human-models/models';
 const modelsOut = 'models/models.json';
+const modelsFolders = [
+  './models',
+  '../human-models/models',
+  '../blazepose/model/',
+  '../anti-spoofing/model',
+  '../efficientpose/models',
+  '../insightface/models',
+  '../movenet/models',
+  '../nanodet/models',
+];
 
 const apiExtractorIgnoreList = [ // eslint-disable-line @typescript-eslint/no-unused-vars
   'ae-missing-release-tag',
@@ -27,29 +37,33 @@ function copy(src, dst) {
 }
 
 async function analyzeModels() {
-  log.info('Analyze:', { modelsDir, modelsOut });
+  log.info('Analyze models:', { folders: modelsFolders.length, result: modelsOut });
   let totalSize = 0;
   const models = {};
-  let dir;
-  try {
-    dir = fs.readdirSync(modelsDir);
-  } catch {
-    log.warn('Cannot enumerate:', modelsDir);
+  const allModels = [];
+  for (const folder of modelsFolders) {
+    try {
+      if (!fs.existsSync(folder)) continue;
+      const stat = fs.statSync(folder);
+      if (!stat.isDirectory) continue;
+      const dir = fs.readdirSync(folder);
+      const found = dir.map((f) => `file://${folder}/${f}`).filter((f) => f.endsWith('json'));
+      log.state('Models', { folder, models: found.length });
+      allModels.push(...found);
+    } catch {
+      // log.warn('Cannot enumerate:', modelFolder);
+    }
   }
-  if (!dir || dir.length === 0) {
-    log.warn('No models found:', modelsDir);
-    return;
-  }
-  for (const f of dir) {
-    if (!f.endsWith('.json')) continue;
-    const url = `file://${modelsDir}/${f}`;
+  for (const url of allModels) {
+    // if (!f.endsWith('.json')) continue;
+    // const url = `file://${modelsDir}/${f}`;
     const model = new tf.GraphModel(url); // create model prototype and decide if load from cache or from original modelurl
     model.findIOHandler();
     const artifacts = await model.handler.load();
     const size = artifacts?.weightData?.byteLength || 0;
     totalSize += size;
-    const name = f.replace('.json', '');
-    models[name] = size;
+    const name = path.basename(url).replace('.json', '');
+    if (!models[name]) models[name] = size;
   }
   const json = JSON.stringify(models, null, 2);
   fs.writeFileSync(modelsOut, json);