+++ test/test-node-gpu.js
@@ -1,5 +1,5 @@
 process.env.TF_CPP_MIN_LOG_LEVEL = '2';
-const Human = require('../dist/human.node-gpu.js').default;
+const H = require('../dist/human.node-gpu.js');
 const test = require('./test-main.js').test;
 
 const config = {
@@ -25,4 +25,8 @@ const config = {
   filter: { enabled: false },
 };
 
-test(Human, config);
+async function main() {
+  test(H.Human, config);
+}
+
+main();