+++ demo/nodejs/node-event.js
@@ -8,7 +8,6 @@ const process = require('process');
 
 let fetch; // fetch is dynamically imported later
 
-// eslint-disable-next-line import/no-extraneous-dependencies, no-unused-vars, @typescript-eslint/no-unused-vars
 const tf = require('@tensorflow/tfjs-node'); // in nodejs environments tfjs-node is required to be loaded before human
 // const human = require('@vladmandic/human'); // use this when human is installed as module (majority of use cases)
 const Human = require('../../dist/human.node.js'); // use this when using human in dev mode
@@ -38,7 +37,7 @@ async function detect(input) {
   let buffer;
   log.info('Loading image:', input);
   if (input.startsWith('http:') || input.startsWith('https:')) {
-    fetch = (await import('node-fetch')).default;
+    fetch = (await import('node-fetch')).default; // eslint-disable-line node/no-extraneous-require, node/no-missing-import
     const res = await fetch(input);
     if (res && res.ok) buffer = await res.buffer();
     else log.error('Invalid image URL:', input, res.status, res.statusText, res.headers.get('content-type'));
@@ -59,6 +58,7 @@ async function main() {
   log.header();
 
   human = new Human.Human(myConfig);
+  log.info('Human:', human.version, 'TF:', tf.version_core);
 
   if (human.events) {
     human.events.addEventListener('warmup', () => {