+++ test/test-node-wasm.js
@@ -1,7 +1,7 @@
 const log = require('@vladmandic/pilogger');
 const tf = require('@tensorflow/tfjs'); // wasm backend requires tfjs to be loaded first
 const wasm = require('@tensorflow/tfjs-backend-wasm'); // wasm backend does not get auto-loaded in nodejs
-const { Canvas, Image } = require('canvas');
+const { Canvas, Image } = require('canvas'); // eslint-disable-line node/no-extraneous-require, node/no-missing-require
 const H = require('../dist/human.node-wasm.js');
 const test = require('./test-main.js').test;