+++ CHANGELOG.md
@@ -11,9 +11,8 @@
   
 ### **HEAD -> main** 2022/08/12 mandic00@live.com
 
-
-### **origin/main** 2022/08/10 mandic00@live.com
-
+- enumerate additional models
+- release refresh
 
 ### **2.9.3** 2022/08/10 mandic00@live.com